moved experiment controls. fixes #3503
