+++ frontend/src/app/components/Warning/index.scss
@@ -48,3 +48,8 @@
     padding: 0;
   }
 }
+
+.warning {
+  margin-bottom: 0;
+  padding: 0;
+}
