+++ frontend/src/app/containers/ExperimentPage/DetailsOverview.js
@@ -5,6 +5,7 @@ import { Localized } from "fluent-react/compat";
 import LocalizedHtml from "../../components/LocalizedHtml";
 import { experimentL10nId, formatDate } from "../../lib/utils";
 import MeasurementsSection from "../../components/Measurements";
+import ExperimentControls from "./ExperimentControls";
 
 import type {
   DetailsOverviewType,
@@ -14,21 +15,58 @@ import type {
 } from "./types";
 
 export default function DetailsOverview({
-  experiment,
+  sendToGA,
+  userAgent,
+  hasAddon,
+  progressButtonWidth,
+  isDisabling,
+  isEnabling,
+  enabled,
+  installed,
   graduated,
+  experiment,
+  installExperiment,
+  doShowEolDialog,
+  doShowPreFeedbackDialog,
+  uninstallExperimentWithSurvey,
   highlightMeasurementPanel,
+  flashMeasurementPanel,
   doShowTourDialog,
-  sendToGA
+  surveyURL
 }: DetailsOverviewType) {
   const { slug, thumbnail, measurements } = experiment;
   const l10nId = (pieces: string) => experimentL10nId(experiment, pieces);
 
   return (
     <div className="details-overview">
-      <div
-        className={`experiment-icon-wrapper-${slug} experiment-icon-wrapper`}
-      >
-        <img className="experiment-icon" src={thumbnail} />
+      <div>
+        <div
+          className={`experiment-icon-wrapper-${slug} experiment-icon-wrapper`}
+        >
+          <img className="experiment-icon" src={thumbnail} />
+        </div>
+        <div>
+          <ExperimentControls
+            {...{
+              sendToGA,
+              hasAddon,
+              userAgent,
+              experiment,
+              installed,
+              graduated,
+              enabled,
+              progressButtonWidth,
+              installExperiment,
+              isEnabling,
+              isDisabling,
+              doShowEolDialog,
+              doShowPreFeedbackDialog,
+              flashMeasurementPanel,
+              uninstallExperimentWithSurvey,
+              surveyURL
+            }}
+          />
+        </div>
       </div>
       <div className="details-sections">
         <section className="user-count">
@@ -82,65 +120,48 @@ export const StatsSection = ({
   },
   sendToGA
 }: StatsSectionType) =>
-  <div>
-    <section className="stats-section">
+  <section className="stats-section">
+    <ul>
       {!web_url &&
-        <p>
-          <Localized id="tourLink">
-            <a className="showTour" onClick={evt => {
-              sendToGA("event", {
-                eventCategory: "ExperimentDetailsPage Interactions",
-                eventAction: "button click",
-                eventLabel: "take tour"
-              });
-              doShowTourDialog(evt);
-            }} href="#">
-              Launch Tour
-            </a>
-          </Localized>
-        </p>}
-      <dl>
-        {changelog_url &&
-          <Localized id="changelog">
-            <dt>Changelog</dt>
-          </Localized>}
-        {changelog_url &&
-          <dd>
+          <li>
+            <Localized id="tourLink">
+              <a className="showTour" onClick={evt => {
+                sendToGA("event", {
+                  eventCategory: "ExperimentDetailsPage Interactions",
+                  eventAction: "button click",
+                  eventLabel: "take tour"
+                });
+                doShowTourDialog(evt);
+              }} href="#">
+                Launch Tour
+              </a>
+            </Localized>
+          </li>}
+      {changelog_url &&
+          <li>
             <a href={changelog_url}>
-              {changelog_url}
+              <Localized id="changelog"><span>Changelog</span></Localized>
             </a>
-          </dd>}
-        <Localized id="contribute">
-          <dt>Contribute</dt>
-        </Localized>
-        <dd>
-          <a href={contribute_url}>
-            {contribute_url}
-          </a>
-        </dd>
-
-        <Localized id="bugReports">
-          <dt>Bug Reports</dt>
-        </Localized>
-        <dd>
-          <a href={bug_report_url}>
-            {bug_report_url}
-          </a>
-        </dd>
-
-        <Localized id="discussExperiment" $title={title}>
-          <dt>
-            Discuss {title}
-          </dt>
-        </Localized>
-        <dd>
-          <a href={discourse_url}>
-            {discourse_url}
-          </a>
-        </dd>
-      </dl>
-    </section>
-  </div>;
+          </li>}
+      <li>
+        <a href={contribute_url}>
+          <Localized id="contribute"><span>Contribute</span></Localized>
+        </a>
+      </li>
+      <li>
+        <a href={bug_report_url}>
+          <Localized id="bugReports"><span>Bug Reports</span></Localized>
+        </a>
+      </li>
+      <li>
+        <a href={discourse_url}>
+          <Localized id="discussExperiment" $title={title}>
+            <span>Discuss {title}</span>
+          </Localized>
+        </a>
+      </li>
+    </ul>
+  </section>;
 
 export const ContributorsSection = ({
   experiment: { contributors, contributors_extra, contributors_extra_url },
