+++ frontend/src/app/containers/ExperimentPage/TestpilotPromo.js
@@ -14,7 +14,7 @@ export default class TestpilotPromo extends React.Component {
   props: TestpilotPromoProps;
 
   render() {
-    const { hasAddon, graduated, experiment, installCallback } = this.props;
+    const { hasAddon, graduated, experiment } = this.props;
 
     const { title, web_url } = experiment;
     if (hasAddon === null || hasAddon || graduated || web_url) {
@@ -39,7 +39,6 @@ export default class TestpilotPromo extends React.Component {
               <MainInstallButton
                 {...this.props}
                 experimentTitle={title}
-                installCallback={installCallback}
               />
             </div>
             <Copter />
