+++ frontend/src/app/containers/ExperimentPage/index.scss
@@ -4,46 +4,35 @@
 .details-header-wrapper {
 
   &:not(.has-status) {
-    padding-top: $grid-unit * 1.8;
-
     .status-bar {
       display: none;
     }
   }
 
+  & > .experiment-icon-wrapper {
+    border-radius: 0;
+    height: 100px;
+  }
+
   @include respond-to('not-small') {
-    &.stick {
-      background: $transparent-white-96;
-      box-shadow: 0 0 5px $transparent-black-5;
-      left: 0;
-      padding-top: 0;
-      position: fixed;
-      top: 0;
-      width: 100%;
-      z-index: 999;
+    &:not(.has-status) {
+      .status-bar {
+        display: block;
+        visibility: hidden;
+      }
+    }
+
+    & > .experiment-icon-wrapper {
+      display: none;
     }
   }
 }
 
 @include respond-to('small') {
 
-  .details-header {
-    padding-top: $grid-unit * .75;
-  }
-
-  .experiment-controls {
-    @include flex-container (column, center, stretch);
-
-    .button {
-      margin: $grid-unit * .5 0;
-      width: 100%;
-    }
-  }
-
-  .highlight-privacy {
+  .details-overview .experiment-icon-wrapper {
     display: none;
   }
-
 
   .sticky-header-sibling {
     display: none;
@@ -54,7 +43,6 @@
   @include flex-container(row, center, center);
   // animation: fade-in 500ms forwards;
   height: $grid-unit * 1.6;
-  margin-bottom: $grid-unit * .2;
   opacity: 0;
   padding-top: $grid-unit * .1;
 
@@ -87,30 +75,38 @@
   margin-bottom: $grid-unit;
 }
 
-.experiment-controls {
-  @include respond-to('not-small') {
-    @include flex-container(row, flex-end, center);
-  }
-
-  flex: 1;
-  position: relative;
-  transition: top 200ms;
-  width: 100%;
-
-  .button {
-    @include respond-to('not-small') {
-      margin-left: $grid-unit;
-    }
-    border-radius: 3px;
-    height: 48px;
-    max-width: 180px;
-    padding: 0 42px;
+.details-controls {
+  @include respond-to('small') {
+    margin-bottom: 0;
   }
 
   .highlight-privacy {
     cursor: pointer;
-    font-size: 15px;
-    font-weight: $weight-semi-bold;
+  }
+}
+
+.experiment-controls {
+  @include respond-to('small') {
+    border: 0;
+    padding: 0;
+  }
+
+  align-items: center;
+  border: 1px solid $grey-20;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 10px;
+
+  .button {
+    border-radius: 3px;
+    height: 48px;
+    margin-top: 8px;
+    width: 100%;
+  }
+
+  .button:first-child {
+    margin-top: 0;
   }
 
   span {
@@ -118,24 +114,24 @@
   }
 }
 
-.details-header {
-  @include respond-to('not-small') {
-    height: $grid-unit * 4.5;
-    margin-bottom: 0;
+.privacy-link {
+  @include respond-to('small') {
+    background: $white;
+    padding-bottom: 0;
+    padding-top: 22px;
   }
+  background: $grey-20;
+  border-radius: 0 0 3px 3px;
+  font-size: 15px;
+  padding: 10px;
+  text-align: center;
+}
 
-  margin-bottom: $grid-unit;
-  position: relative;
-
+.details-header {
   h1 {
-    @include respond-to('small') {
-      margin: 15px 0;
-    }
-
     display: flex;
     flex: 1;
     font-weight: $weight-light;
-    margin: 0;
   }
 
   h4 {
@@ -181,9 +177,7 @@
 
   padding-top: $grid-unit * .5;
 
-  .contributors-section,
-  .stats-section,
-  .measurements {
+  .details-sections {
     @include hide-on-small;
   }
 }
@@ -281,10 +275,10 @@
 }
 
 .details-sections {
+  margin-top: 20px;
 
   section {
     margin-bottom: $grid-unit;
-    padding-bottom: $grid-unit;
 
     h3 {
       font-size: 22px;
@@ -297,10 +291,8 @@
 }
 
 .user-count {
-  background: $transparent-black-05;
-  font-size: 15px;
-  padding: $grid-unit;
-  text-align: center;
+  font-size: 22px;
+  font-weight: $weight-light;
 }
 
 .disclaimer,
@@ -387,32 +379,8 @@
 .stats-section {
   width: 100%;
 
-  dl {
-    margin-bottom: 0;
-  }
-
-  dt {
-    font-size: 15px;
-    font-weight: regular;
-    margin: 0 0 $grid-unit * .25;
-  }
-
-  dd {
-    font-size: 13px;
-    font-weight: regular;
-    margin: 0 0 12px;
-    width: 100%;
-    word-break: break-all;
-  }
-
-  dd:last-child {
-    margin-bottom: 0;
-  }
-
-  p {
-    font-size: 15px;
-    font-weight: $weight-semi-bold;
-    width: 35%;
+  ul {
+    padding-left: 1em;
   }
 }
